+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
-## 9/13/2017
-- Added ability to read binary `.ply` files.
-- Added support for `.csv` files with or without `SCAN_INDEX` and `SIGNAL_STRENGTH` columns.
+## 7/13/2021
+- Added support for RGB .csv and .ply file.
+- Added function of coordinates offset correction.
+- Added function of transform from top-down scale to room scale.